Adding ConfigurationError exception
