+++ sam-hiberry-app/src/integration/shopify_processor/exceptions.py
@@ -1,0 +1,5 @@
+
+class ConfigurationError(Exception):
+    """Exception raised for errors in the configuration."""
+    pass
+
